requirements.txt
accelerate==1.2.1
adabelief-pytorch==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
bm3d==4.0.3
bm4d==4.2.5
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
diffusers==0.20.2
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
expecttest==0.2.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.25.0
hypothesis==6.103.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonpointer==2.1
kiwisolver==1.4.7
lark==1.1.9
lazy_loader==0.4
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.0
nvitop==1.3.2
ollama==0.4.4
opencv-python==4.10.0.84
optree==0.11.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
reedsolo==1.7.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.1
wcwidth==0.2.13
x-transformers==1.29.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0